adds triggerUpdate function, starts liveUpdate effect
added triggerUpdate function to update isUpdating property of reports
state to true. This is so that this state can change at the proper
intervals for the corresponding result. This function is prop drilled
through ResultsList to Result component.

Began work on liveUpdate effect which is intended to start an interval
every 5 minutes that interates through reports and runs function to update
each report that is queued up to update.